requirements.txt
openai>=1.3.0
google-genai>=0.5.0
gradio>=4.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.10
easyocr